✅ S2TV0X task: clear v0.5 release gates
