+++ scripts/oversized-test-baseline.json
@@ -3,13 +3,13 @@
   "threshold_lines": 1000,
   "budgets": {
     "max_entries": 10,
-    "max_total_lines": 11429,
+    "max_total_lines": 11472,
     "max_file_lines": 1300
   },
   "entries": [
     {
       "file": "packages/agentplane/src/cli/run-cli.core.incidents.test.ts",
-      "lines": 1112
+      "lines": 1152
     },
     {
       "file": "packages/agentplane/src/cli/run-cli.core.pr-flow.pr-validation.test.ts",
@@ -29,7 +29,7 @@
     },
     {
       "file": "packages/agentplane/src/cli/run-cli.core.tasks.query-listing.test.ts",
-      "lines": 1190
+      "lines": 1193
     },
     {
       "file": "packages/agentplane/src/cli/run-cli.core.test.ts",
